# src/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UE API",
    "description": "This is a Classification API application made with Express and documented with Swagger",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/projects": {
      "post": {
        "summary": "Add a new project",
        "tags": ["Projects"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "example": {
                  "message": "Project created successfully",
                  "project": {
                    "projectName": "My Project",
                    "description": "This is a sample project",
                    "discipline": "Computer Science",
                    "type": "Web application",
                    "authors": ["Author1", "Author2"],
                    "yearOfSubmission": 2022
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Validation error.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Validation error details"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all projects",
        "tags": ["Projects"],
        "responses": {
          "200": {
            "description": "Returns a list of projects",
            "content": {
              "application/json": {
                "example": {
                  "projects": [
                    {
                      "projectName": "Project1",
                      "description": "Description1",
                      "discipline": "Discipline1",
                      "type": "Type1",
                      "authors": ["Author1"],
                      "yearOfSubmission": 2022
                    },
                    {
                      "projectName": "Project2",
                      "description": "Description2",
                      "discipline": "Discipline2",
                      "type": "Type2",
                      "authors": ["Author2"],
                      "yearOfSubmission": 2022
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
  },
  "/api/projects/discipline/{discipline}": {
    "get": {
      "summary": "Get projects by discipline",
      "tags": ["Projects"],
      "parameters": [
        {
          "in": "path",
          "name": "discipline",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Discipline of the projects to retrieve"
        }
      ],
      "responses": {
        "200": {
          "description": "List of projects by discipline",
          "content": {
            "application/json": {
              "example": {
                "projects": [
                  {
                    "projectName": "Project1",
                    "description": "Description1",
                    "discipline": "Discipline1",
                    "type": "Type1",
                    "authors": ["Author1"],
                    "yearOfSubmission": 2022
                  },
                  {
                    "projectName": "Project2",
                    "description": "Description2",
                    "discipline": "Discipline1",
                    "type": "Type2",
                    "authors": ["Author2"],
                    "yearOfSubmission": 2022
                  }
                ]
              }
            }
          }
        },
        "404": {
          "description": "No projects found for the specified discipline",
          "content": {
            "application/json": {
              "example": {
                "message": "No projects found for the specified discipline"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/type/{type}": {
    "get": {
      "summary": "Get projects by type",
      "tags": ["Projects"],
      "parameters": [
        {
          "in": "path",
          "name": "type",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Type of the projects to retrieve"
        }
      ],
      "responses": {
        "200": {
          "description": "List of projects by type",
          "content": {
            "application/json": {
              "example": {
                "projects": [
                  {
                    "projectName": "Project1",
                    "description": "Description1",
                    "discipline": "Discipline1",
                    "type": "Type1",
                    "authors": ["Author1"],
                    "yearOfSubmission": 2022
                  },
                  {
                    "projectName": "Project2",
                    "description": "Description2",
                    "discipline": "Discipline2",
                    "type": "Type1",
                    "authors": ["Author2"],
                    "yearOfSubmission": 2022
                  }
                ]
              }
            }
          }
        },
        "404": {
          "description": "No projects found for the specified type",
          "content": {
            "application/json": {
              "example": {
                "message": "No projects found for the specified type"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/discipline/{discipline}/type/{type}": {
    "get": {
      "summary": "Get projects by discipline and type",
      "tags": ["Projects"],
      "parameters": [
        {
          "in": "path",
          "name": "discipline",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Discipline of the projects to retrieve"
        },
        {
          "in": "path",
          "name": "type",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Type of the projects to retrieve"
        }
      ],
      "responses": {
        "200": {
          "description": "List of projects by discipline and type",
          "content": {
            "application/json": {
              "example": {
                "projects": [
                  {
                    "projectName": "Project1",
                    "description": "Description1",
                    "discipline": "Discipline1",
                    "type": "Type1",
                    "authors": ["Author1"],
                    "yearOfSubmission": 2022
                  },
                  {
                    "projectName": "Project2",
                    "description": "Description2",
                    "discipline": "Discipline1",
                    "type": "Type2",
                    "authors": ["Author2"],
                    "yearOfSubmission": 2022
                  }
                ]
              }
            }
          }
        },
        "404": {
          "description": "No projects found for the specified discipline and type",
          "content": {
            "application/json": {
              "example": {
                "message": "No projects found for the specified discipline and type"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/search": {
    "get": {
      "summary": "Search projects",
      "tags": ["Projects"],
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectSearchQuery"
        }
      ],
      "responses": {
        "200": {
          "description": "List of projects matching the search query",
          "content": {
            "application/json": {
              "example": {
                "projects": [
                  {
                    "projectName": "Project1",
                    "description": "Description1",
                    "discipline": "Discipline1",
                    "type": "Type1",
                    "authors": ["Author1"],
                    "yearOfSubmission": 2022
                  },
                  {
                    "projectName": "Project2",
                    "description": "Description2",
                    "discipline": "Discipline2",
                    "type": "Type2",
                    "authors": ["Author2"],
                    "yearOfSubmission": 2022
                  }
                ]
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/id/{id}": {
    "get": {
      "summary": "Get project by ID",
      "tags": ["Projects"],
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "responses": {
        "200": {
          "description": "Project details",
          "content": {
            "application/json": {
              "example": {
                "projectName": "Project1",
                "description": "Description1",
                "discipline": "Discipline1",
                "type": "Type1",
                "authors": ["Author1"],
                "yearOfSubmission": 2022
              }
            }
          }
        },
        "404": {
          "description": "Project not found",
          "content": {
            "application/json": {
              "example": {
                "message": "Project not found"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/{id}": {
    "put": {
      "summary": "Update a project by ID",
      "tags": ["Projects"],
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "ID of the project to update"
        }
      ],
      "requestBody": {
        "required": true,
        "content": {
          "multipart/form-data": {
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Project updated successfully",
          "content": {
            "application/json": {
              "example": {
                "message": "Project updated successfully",
                "project": {
                  "projectName": "Updated Project",
                  "description": "Updated project description",
                  "discipline": "Computer Science",
                  "type": "Web application",
                  "authors": ["UpdatedAuthor"],
                  "yearOfSubmission": 2022
                }
              }
            }
          }
        },
        "400": {
          "description": "Bad request. Validation error.",
          "content": {
            "application/json": {
              "example": {
                "message": "Validation error details"
              }
            }
          }
        },
        "404": {
          "description": "Project not found",
          "content": {
            "application/json": {
              "example": {
                "message": "Project not found"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "/api/projects/delete/{id}": {
    "put": {
      "summary": "Delete a project by ID",
      "tags": ["Projects"],
      "security": [
        {
          "bearerAuth": []
        }
      ],
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "schema": {
            "type": "string"
          },
          "required": true,
          "description": "ID of the project to delete"
        }
      ],
      "responses": {
        "200": {
          "description": "Project deleted successfully",
          "content": {
            "application/json": {
              "example": {
                "message": "Project deleted successfully"
              }
            }
          }
        },
        "404": {
          "description": "Project not found",
          "content": {
            "application/json": {
              "example": {
                "message": "Project not found"
              }
            }
          }
        },
        "500": {
          "description": "Internal server error",
          "content": {
            "application/json": {
              "example": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  }
},
...
}